feat: simplify setup and teardown
Let the end-user specify the Prisma connection URL directly with an option
to configure the envar name of the connection URL.

Remove all specific adapters, and instead utilize the built-in Prisma CLI
commands for bootstrapping the DB and teardown of the DB.

Drop peer-deps for PostgreSQL and MySQL client packages.

Update the README to document the breaking changes and changes in underlying
behaviour.

BREAKING CHANGE: change environment options and remove creds envars